+++ lib/src/app_version.dart
@@ -1,3 +1,3 @@
-const String kMindVersionLabel = 'v0.3.12';
+const String kMindVersionLabel = 'v0.3.13';
 const String kMindChangelogUrl =
     'https://github.com/jakob37/mind/blob/master/CHANGELOG.md';